BOL-35 feat: GroupRepository 추가
