+++ adapter-out/rdb/src/main/kotlin/com/yapp/bol/group/GroupCommandRepositoryImpl.kt
@@ -1,5 +1,7 @@
 package com.yapp.bol.group
 
+import com.yapp.bol.group.member.Member
+import com.yapp.bol.group.member.MemberList
 import com.yapp.bol.group.member.toEntity
 import org.springframework.stereotype.Repository
 
@@ -13,3 +15,34 @@ internal class GroupCommandRepositoryImpl(
         return groupEntity.toDomain()
     }
 }
+
+internal fun Group.toEntity(): GroupEntity = GroupEntity(
+    id = id,
+    name = name,
+    description = description,
+    organization = organization,
+    profileImageUrl = profileImageUrl,
+    accessCode = accessCode,
+    members = members.toList().map {
+        it.toEntity()
+    }
+)
+
+internal fun GroupEntity.toDomain(): Group = Group(
+    id = id,
+    name = name,
+    description = description,
+    organization = organization,
+    profileImageUrl = profileImageUrl,
+    accessCode = accessCode,
+    members = members.map {
+        Member(
+            id = it.id,
+            userId = it.userId,
+            role = it.role,
+            nickname = it.nickname,
+        )
+    }.let {
+        MemberList(it.toMutableList())
+    }
+)
